Rewrote Knife Juggler to correctly work with upstream, wrote unit tests for juggles
Also, wrote it such that juggles' throw is a random effect node that comes after a minion being placed (with each child being juggles hitting a different target),
so that the AI can correctly look at all possible throw targets and see how they stack up (rather than how we had it, which was that juggles threw without generating any new nodes).

In theory, this means we can log juggles' throw, but probably not in practice.