New logger!
Logging now works with copies of cards, right as the event that we want to log happens.  This slows down the sim _a ton_ (from about 0.2 seconds a game to 1 sec a game), but
helps me keep my sanity (there was an indexing edge case with correctly logging top decking, because indexes for cards get stored before an action, but state got stored after an action).

Logging is in for attacks and minion placement, with the targets for minion placement getting weird.  I'm gonna fix that soon.  Hero Powers should also be a very easy addition.

This is a WIP branch, and as such needs sanity checking and testing.